refactoring
- Extrat commit graph traversal logic out of CommitMetrics's constructor
- renaming and simplified logic
