+++ src/info/author.rs
@@ -1,6 +1,7 @@
+use super::git::metrics::GitMetrics;
 use crate::{
     cli::NumberSeparator,
-    info::{format_number, utils::git::CommitMetrics, utils::info_field::InfoField},
+    info::{format_number, utils::info_field::InfoField},
 };
 use owo_colors::{DynColors, OwoColorize};
 use serde::Serialize;
@@ -68,8 +69,8 @@ pub struct AuthorsInfo {
 }
 
 impl AuthorsInfo {
-    pub fn new(info_color: DynColors, commit_metrics: &CommitMetrics) -> Self {
-        let authors = commit_metrics.authors_to_display.clone();
+    pub fn new(info_color: DynColors, git_metrics: &GitMetrics) -> Self {
+        let authors = git_metrics.authors_to_display.clone();
         Self {
             authors,
             info_color,
