+++ src/info/contributors.rs
@@ -1,9 +1,9 @@
-use serde::Serialize;
-
+use super::git::metrics::GitMetrics;
 use crate::{
     cli::NumberSeparator,
-    info::{format_number, utils::git::CommitMetrics, utils::info_field::InfoField},
+    info::{format_number, utils::info_field::InfoField},
 };
+use serde::Serialize;
 
 #[derive(Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -17,12 +17,12 @@ pub struct ContributorsInfo {
 
 impl ContributorsInfo {
     pub fn new(
-        commit_metrics: &CommitMetrics,
+        git_metrics: &GitMetrics,
         number_of_authors_to_display: usize,
         number_separator: NumberSeparator,
     ) -> Self {
         Self {
-            total_number_of_authors: commit_metrics.total_number_of_authors,
+            total_number_of_authors: git_metrics.total_number_of_authors,
             number_of_authors_to_display,
             number_separator,
         }
@@ -47,14 +47,12 @@ impl InfoField for ContributorsInfo {
 #[cfg(test)]
 mod test {
     use super::*;
+    use gix::actor::Time;
 
     #[test]
     fn test_display_contributors_info() {
-        use crate::info::utils::git::CommitMetrics;
-        use gix::actor::Time;
-
         let timestamp = Time::now_utc();
-        let commit_metrics = CommitMetrics {
+        let git_metrics = GitMetrics {
             authors_to_display: vec![],
             file_churns_to_display: vec![],
             total_number_of_authors: 12,
@@ -65,7 +63,7 @@ mod test {
             time_of_first_commit: timestamp,
         };
 
-        let contributors_info = ContributorsInfo::new(&commit_metrics, 2, NumberSeparator::Plain);
+        let contributors_info = ContributorsInfo::new(&git_metrics, 2, NumberSeparator::Plain);
         assert_eq!(contributors_info.value(), "12".to_string());
         assert_eq!(contributors_info.title(), "Contributors".to_string());
     }
